requirements.txt
python-telegram-bot>=20.0
pyrogram>=2.0.106
TgCrypto>=1.2.5
httpx>=0.27